requirements.txt
pandas
numpy
azure-storage-blob
